Refactor: Decompose monolithic app.js into smaller modules
This refactoring breaks down the monolithic `js/app.js` file into a more organized and maintainable modular structure.

The key changes include:
- A new `js/components/App.js` file now holds the main `ModernHomepage` class.
- New modules under `js/modules/` handle specific concerns: `dom.js`, `state.js`, `events.js`, `theme.js`, and `modals.js`.
- The main `js/app.js` file is now a clean entry point that initializes the application.
- `index.html` has been updated to load `app.js` as a module.

This new structure improves code organization, reduces complexity, and makes future maintenance easier without altering any existing functionality.
